\name{family.hpdglm}
\alias{family.hpdglm}
\title{
    family of an hpdglm model
}
\description{
    Returns the family used for building an hpdglm model.
}
\usage{
    family.hpdglm(object, ...)
}
\arguments{
  \item{object}{an hpdglm model}
}
\value{the family of the model}

\keyword{ family }
\keyword{ hpdglm }
