\name{HPdregression-package}
\alias{HPdregression-package}
\alias{HPdregression}
\docType{package}
\title{ HPdregression - Distributed Regression for Big Data}
\description{
  \pkg{HPdregression} provides distributed algorithms for regression models. It is written based on the infrastructure created in HP-Lab for distributed computing in R.
}
\details{
 \tabular{ll}{
  Package: \tab HPdregression\cr
  Type: \tab Package\cr
  Version: \tab 1.1.0\cr
  Date: \tab 2015-01-16\cr
 }
 Main Functions:
 \itemize{
  \item {hpdglm:} {It is a distributed version of glm.}
  \item {v.hpdglm:} {This function is implemented for evaluating a model built by hpdglm using Split-Sample-Validation method.}
  \item {cv.hpdglm:} {This function is implemented for evaluating a model built by hpdglm using Cross-Validation method.}
 }

}

\author{
HP Vertica Analytics Team <distributedRTeam@external.groups.hp.com>
}

\keyword{distributed R, distributed Regression, Big Data Analytics}


